requirements.txt
fastapi==0.115.2
uvicorn==0.34.0
SQLAlchemy==2.0.37
GeoAlchemy2==0.17.0
asyncpg==0.30.0
alembic==1.14.1
python-dotenv==1.0.1
psycopg2-binary==2.9.10